Clean up ARM_TOOLCHAIN
Rename `ARM_TOOCHAIN` to `ARM_TOOLCHAIN` and use a sensible default.
On Ubuntu (and possibly other Debian-like distros), the ARM toolchain
can be installed with the distro's package manager in `/usr/bin`,
so it makes more sense to use that as a default instead of some very
specific path name in `/opt`.