Add configuration files for VS Code
Uses Bear to create a compilation database from the `make` run so that
the C/C++ Intellisense will navigate to the correct definitions.

This build task is run before launching the pyOCD or J-Link debugger.

Note that when changing between targets, you may have to restart VS Code
so that the compilation database is reloaded. I have not found a way to
trigger this from the build task.