rules.mk: 'flash' command now builds the required .elf file
Pyocd is able to flash .elf files use that to flash
Running 'make clean && make flash' now works as expected

Only tested the .elf file for pyocd for now.
Jlink command does not directly reference .elf/.bin/.hex. If it needs .elf, this works for both, otherwises it remains the same for Jlink