Allow use as a git submodule
Reorganize the Makefile to put most of the device-specific paths
into `rules.mk`.

The paths are handled so that py32f0-template may be used as a git
submodule within another repository. The following structures are
possible:

    single/
        Build/  (generated)
        py32f0-template/  (submodule)
        Makefile  (adapted from py32f0-template/Makefile by user)
        main.c

or if you have multiple builds in your repository:

    multi/
        a/
            Build/  (generated)
            Makefile  (adapted from ../py32f0-template/Makefile)
            main.c
        b/
            Build/  (generated)
            Makefile  (adapted from ../py32f0-template/Makefile)
            main.c
        py32f0-template/  (submodule)

Of course the repository can still be used stand-alone as before.

Additionally, allow the use of *.S assembler files. The uppercase
extension means that gcc will run it through the preprocessor.